backend/services/image-processing/requirements.txt
rembg==2.0.50
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
